chore(back): Merge Explorer and Annotator apps
