+++ pixano/app/README.md
@@ -11,35 +11,33 @@
 
 </div>
 
-# Pixano Explorer
-
-## Launching the Explorer
+## Launching the Pixano app
 
 ### From a terminal
 
-You can start the Explorer app with the following command:
+You can start the Pixano app with the following command:
 
 ```shell
-pixano-explorer <path/to/your/datasets>
+pixano <path/to/your/datasets>
 ```
 
 You will then be provided with a URL to open in your browser to use the app.
 
 ### From a notebook
 
-If you are in a Jupyter or Google Colab notebook, you can start the Explorer app by running a cell with:
+If you are in a Jupyter or Google Colab notebook, you can start the app by running a cell with:
 
 ```python
-from pixano.apps import Explorer
-explorer = Explorer(<path/to/your/datasets>)
+from pixano.app import App
+app = App(<path/to/your/datasets>)
 ```
 
 You can then use the app directly from the notebook in another cell with:
 
 ```python
-explorer.display()
+app.display()
 ```
 
-## Using the Explorer
+## Using the Pixano app
 
-Please refer to our [user guide for Pixano Explorer](https://pixano.github.io/user/explorer/) on our documentation website for more information on how to use it.
+Please refer to our [user guide](https://pixano.github.io/user/app/) on our documentation website for more information on how to use it.
