--- pixano/apps/__init__.py
@@ -1,22 +1,0 @@
-# @Copyright: CEA-LIST/DIASI/SIALV/LVA (2023)
-# @Author: CEA-LIST/DIASI/SIALV/LVA <pixano@cea.fr>
-# @License: CECILL-C
-#
-# This software is a collaborative computer program whose purpose is to
-# generate and explore labeled data for computer vision applications.
-# This software is governed by the CeCILL-C license under French law and
-# abiding by the rules of distribution of free software. You can use,
-# modify and/ or redistribute the software under the terms of the CeCILL-C
-# license as circulated by CEA, CNRS and INRIA at the following URL
-#
-# http://www.cecill.info
-
-from pixano.apps.annotator import Annotator
-from pixano.apps.explorer import Explorer
-from pixano.apps.main import create_app
-
-__all__ = [
-    "Annotator",
-    "Explorer",
-    "create_app",
-]
